@@ -1,37 +1,37 @@
 package jp.supership.nativeads;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.widget.FrameLayout;
-
-import com.facebook.ads.NativeAd;
-
-import com.facebook.ads.NativeBannerAd;
 import com.socdm.d.adgeneration.ADG;
 import com.socdm.d.adgeneration.ADGConsts;
 import com.socdm.d.adgeneration.ADGListener;
 import com.socdm.d.adgeneration.nativead.ADGNativeAd;
-
+import jp.supership.nativeads.databinding.ActivityMainBinding;
 import jp.supership.nativeads.views.nativeads.ADGNativeAdView;
-import jp.supership.nativeads.views.nativeads.FBNativeAdView;
-import jp.supership.nativeads.views.nativeads.FBNativeBannerAdView;
 
 public class MainActivity extends AppCompatActivity {
 
+    private ActivityMainBinding binding;
     private FrameLayout adContainer;
     private ADG adg;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        binding = ActivityMainBinding.inflate(getLayoutInflater());
+        View view = binding.getRoot();
+        setContentView(view);
 
         adg = new ADG(this);
 
         // 管理画面から払い出された広告枠ID
         adg.setLocationId("48636");
+
+        // テストモードを有効化
+        //adg.setEnableTestMode(true);
 
         /**
          * 枠サイズ
@@ -52,15 +52,10 @@ public class MainActivity extends AppCompatActivity {
         adg.setAdListener(new AdListener());
 
         // HTMLテンプレートを使用したネイティブ広告を表示のためにはaddViewする必要があります
-        adContainer = (FrameLayout) findViewById(R.id.ad_container);
+        adContainer = binding.adContainer;
         adContainer.addView(adg);
 
-        /**
-         * 実機でAudience Networkのテスト広告を表示する場合、
-         * ログに出力されるデバイスハッシュを取得し、addTestDeviceを実行してください
-         * AdSettings.addTestDevice("{device_hash}");
-         */
-        findViewById(R.id.btn_stopstart).setOnClickListener(new View.OnClickListener() {
+        binding.btnStopstart.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 if (adg != null) {
@@ -97,14 +92,6 @@ public class MainActivity extends AppCompatActivity {
             if (o instanceof ADGNativeAd) {
                 ADGNativeAdView nativeAdView = new ADGNativeAdView(MainActivity.this);
                 nativeAdView.apply((ADGNativeAd) o);
-                view = nativeAdView;
-            } else if (o instanceof NativeAd) {
-                FBNativeAdView nativeAdView = new FBNativeAdView(getApplicationContext());
-                nativeAdView.apply((NativeAd) o);
-                view = nativeAdView;
-            } else if (o instanceof NativeBannerAd) {
-                FBNativeBannerAdView nativeAdView = new FBNativeBannerAdView(getApplicationContext());
-                nativeAdView.apply((NativeBannerAd) o);
                 view = nativeAdView;
             }
 
